+++ src/colvarbias_meta.h
@@ -232,8 +232,8 @@ protected:
   std::vector<int> reflection_llimit_cv;
   std::vector<int> reflection_ulimit_cv;
   cvm::real reflection_int;
-  int nrefvarsl;
-  int nrefvarsu;
+  size_t nrefvarsl;
+  size_t nrefvarsu;
 
   /// \brief Limits for reflection 
   std::vector<cvm::real> reflection_llimit;
@@ -258,8 +258,8 @@ protected:
 
   std::vector<int> which_int_llimit_cv;
   std::vector<int> which_int_ulimit_cv;
-  int nintvarsl;
-  int nintvarsu;
+  size_t nintvarsl;
+  size_t nintvarsu;
   /// \brief Limits for interval 
   std::vector<cvm::real> interval_llimit;
   std::vector<cvm::real> interval_ulimit;
